% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{Geometric}
\alias{Geometric}
\title{Geometric random variable.}
\usage{
Geometric(k, p)
}
\arguments{
\item{k}{A number.}

\item{p}{A number.}
}
\value{
The probability of having the first success on the \code{k} th attempt with a probability \code{p} of succeeding on any one try.
}
\description{
Geometric random variable.
}
\examples{
Geometric(10, 0.1)
Geometric(99, 0.55)
}
